fix: utils 함수 @lib 내부로 이관
